Added fit with transformed reference spectrum
Can now use fitBranchToRef to fit a branch using a reference spectrum. The parameters of the reference spectrum's model that correspond to locations on the x axis are allowed to vary only in unison via a single set of transformation parameters. The rest of the parameters in the reference model are allowed to vary freely.

The optimal values and covariance for the transformation parameters represent a usable transformation from values on the x axis of the dataset being fit to values on the x axis of the dataset that made the reference spectrum.