Pre-implementation of expanded cut functionality
Added arguments to dev-analyze.py that will be used to specify the planned expansion to cutting functionality.

This functionality will include:
Ability to cut on polygons by specifying  expressions and vertices; points that are in the convex hull of the set of vertices will pass.
Ability to save the results of cuts to a file, including the boolean filter arrays (by event number, which events pass a given cut) and index arrays (array of all indices containing True for the former array.)