Fix inability to fit defined branches
Previously, there was no check that a branch was not a user-expression-defined branch when requesting branches from files, meaning that trying to fit or plot a defined branch would fail.

Implemented a check where branches that are defined as expressions of other branches are excluded from the set of branches requested from files. These branches were not explicitly requested before, but they were requested if they were used in a fit expression.