Add support for trailing AND in cli arguments
AND is used as a delimiter between sets of arguments for defining multiple calls to be plotted on the same axes. Previously, if the arguments after the last instance of AND were not a complete call, the program would error.
This update adds a check, and in the case where the final argument set has no positional arguments, it is merged with the second-to-last set. This allows for the use of trailing AND delimiters, which eases the construction of complex function calls.