Unify single call and multiple call handling
Moved function for splitting cli arguments into multiple calls to the utils.cli module.
Unified handling of single- and multi-call cases in fitBranch.py and dev-analyze.py. Now, we compose a list of calls which has one or more entries, and then handle that list with the same code whether or not there are multiple calls.